api/requirements.txt
flask
flask-cors
python-dotenv
gunicorn
pillow
python-magic
